clean up contracts, make tests configurable with truffle from existing synthetix tests

+++ contracts/Pausable.sol
@@ -4,8 +4,9 @@ pragma solidity ^0.8.0;
 // Inheritance
 import "./Owned.sol";
 
+// https://docs.synthetix.io/contracts/source/contracts/pausable
 abstract contract Pausable is Owned {
-    uint public lastPauseTime;
+    uint256 public lastPauseTime;
     bool public paused;
 
     constructor() internal {
@@ -39,7 +40,10 @@ abstract contract Pausable is Owned {
     event PauseChanged(bool isPaused);
 
     modifier notPaused {
-        require(!paused, "This action cannot be performed while the contract is paused");
+        require(
+            !paused,
+            "This action cannot be performed while the contract is paused"
+        );
         _;
     }
 }
