+++ contracts/TestablePausable.sol
@@ -3,16 +3,15 @@ pragma solidity ^0.8.0;
 import "./Owned.sol";
 import "./Pausable.sol";
 
-
 /**
  * @title An implementation of Pausable. Used to test the features of the Pausable contract that can only be tested by an implementation.
  */
 contract TestablePausable is Owned, Pausable {
-    uint public someValue;
+    uint256 public someValue;
 
     constructor(address _owner) public Owned(_owner) Pausable() {}
 
-    function setSomeValue(uint _value) external notPaused {
+    function setSomeValue(uint256 _value) external notPaused {
         someValue = _value;
     }
 }
